fix: traffic data for other profiles
