+++ openrouteservice/src/main/java/org/heigit/ors/mapmatching/GhMapMatcher.java
@@ -1,6 +1,8 @@
 package org.heigit.ors.mapmatching;
 
 import com.graphhopper.GraphHopper;
+import com.graphhopper.matching.MapMatching;
+import com.graphhopper.matching.MatchResult;
 import com.graphhopper.matching.Observation;
 import com.graphhopper.util.EdgeIteratorState;
 import com.graphhopper.util.PMap;
@@ -9,8 +11,6 @@ import com.graphhopper.util.shapes.GHPoint;
 import org.jetbrains.annotations.NotNull;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Geometry;
-import com.graphhopper.matching.MapMatching;
-import com.graphhopper.matching.MatchResult;
 
 import java.util.Arrays;
 import java.util.List;
@@ -20,16 +20,16 @@ import java.util.List;
  * AbstractMapMatcher using Graphhopper's MapMatching internally.
  */
 public class GhMapMatcher extends AbstractMapMatcher {
-    public GhMapMatcher(GraphHopper graphHopper) {
+    MapMatching mapMatching;
+    public GhMapMatcher(GraphHopper graphHopper, String profile) {
         setGraphHopper(graphHopper);
+        PMap hints = new PMap()
+            .putObject("profile", profile)
+            .putObject(Parameters.Landmark.DISABLE, true);
+        mapMatching = new MapMatching(graphHopper, hints);
     }
     @Override
     public RouteSegmentInfo[] match(Coordinate[] locations, boolean bothDirections) {
-        PMap hints = new PMap()
-                .putObject("profile", "car_ors_fastest")
-                .putObject(Parameters.Landmark.DISABLE, true);
-        MapMatching mapMatching = new MapMatching(graphHopper, hints);
-
         List<Observation> inputGPXEntries = getObservationsFromLocations(locations);
         MatchResult mr = mapMatching.match(inputGPXEntries);
         return getRouteSegmentInfoFromMatchResult(mr);
