+++ openrouteservice/src/main/java/org/heigit/ors/routing/configuration/RoutingManagerConfiguration.java
@@ -15,6 +15,7 @@ package org.heigit.ors.routing.configuration;
 
 import com.graphhopper.util.Helper;
 import com.typesafe.config.ConfigFactory;
+import org.heigit.ors.routing.RoutingProfileType;
 import org.locationtech.jts.geom.Envelope;
 import org.heigit.ors.config.IsochronesServiceSettings;
 import org.heigit.ors.config.RoutingServiceSettings;
@@ -171,6 +172,9 @@ public class RoutingManagerConfiguration  {
 							for(Map.Entry<String, Object> entry : entryValue.entrySet()) {
 								storageParams.put(entry.getKey(), StringUtility.trimQuotes(entry.getValue().toString()));
 							}
+
+							storageParams.put("gh_profile", RoutingProfileType.getEncoderName(RoutingProfileType.getFromString(profile.getProfiles()))+"_fastest" + (RouteProfileConfiguration.hasTurnCosts(profile.getEncoderOptions()) ? "_with_turn_costs" : ""));
+
 
 							profile.getExtStorages().put(storageEntry.getKey(), storageParams);
 						}
