+++ openrouteservice/src/test/resources/ors-config.json
@@ -285,6 +285,15 @@
                   "boundaries": "./src/test/files/data/borders/borders.geojson",
                   "ids": "./src/test/files/data/borders/ids.csv",
                   "openborders": "./src/test/files/data/borders/openborders.csv"
+                },
+                "HereTraffic": {
+                  "enabled": true,
+                  "streets": "./src/test/files/data/traffic_data/here/Streets.shp",
+                  "ref_pattern": "./src/test/files/data/traffic_data/here/ref_pattern.csv",
+                  "pattern_15min": "./src/test/files/data/traffic_data/here/pattern_kph_15min.csv",
+                  "radius": 150,
+                  "output_log": false,
+                  "log_location": "ors/traffic_log"
                 }
               },
               "traffic": false
